Made executor API available in PRO only
