+++ services/src/handlers/plots.rs
@@ -14,7 +14,7 @@ use geoengine_operators::engine::{
 };
 
 use crate::error;
-use crate::error::{Error, Result};
+use crate::error::Result;
 use crate::handlers::Context;
 use crate::ogc::util::{parse_bbox, parse_time};
 use crate::util::parsing::parse_spatial_resolution;
@@ -127,9 +127,11 @@ async fn get_plot_handler<C: Context>(
     session: C::Session,
     ctx: web::Data<C>,
 ) -> Result<impl Responder> {
-    let workflow_id = WorkflowId(id.into_inner());
-
-    let workflow = ctx.workflow_registry_ref().await.load(&workflow_id).await?;
+    let workflow = ctx
+        .workflow_registry_ref()
+        .await
+        .load(&WorkflowId(id.into_inner()))
+        .await?;
 
     let operator = workflow.operator.get_plot().context(error::Operator)?;
 
@@ -167,51 +169,42 @@ async fn get_plot_handler<C: Context>(
     };
 
     let query_ctx = ctx.query_context()?;
-    let task_manager = ctx.task_manager();
 
     let output_format = PlotOutputFormat::from(&processor);
     let plot_type = processor.plot_type();
 
-    let task = async move {
-        Ok(match processor {
-            TypedPlotQueryProcessor::JsonPlain(processor) => processor
+    let data = match processor {
+        TypedPlotQueryProcessor::JsonPlain(processor) => processor
+            .plot_query(query_rect, &query_ctx)
+            .await
+            .context(error::Operator)?,
+        TypedPlotQueryProcessor::JsonVega(processor) => {
+            let chart = processor
                 .plot_query(query_rect, &query_ctx)
                 .await
-                .context(error::Operator)?,
-            TypedPlotQueryProcessor::JsonVega(processor) => {
-                let chart = processor
-                    .plot_query(query_rect, &query_ctx)
-                    .await
-                    .context(error::Operator)?;
+                .context(error::Operator)?;
 
-                serde_json::to_value(&chart).context(error::SerdeJson)?
-            }
-            TypedPlotQueryProcessor::ImagePng(processor) => {
-                let png_bytes = processor
-                    .plot_query(query_rect, &query_ctx)
-                    .await
-                    .context(error::Operator)?;
+            serde_json::to_value(&chart).context(error::SerdeJson)?
+        }
+        TypedPlotQueryProcessor::ImagePng(processor) => {
+            let png_bytes = processor
+                .plot_query(query_rect, &query_ctx)
+                .await
+                .context(error::Operator)?;
 
-                let data_uri = format!("data:image/png;base64,{}", base64::encode(png_bytes));
+            let data_uri = format!("data:image/png;base64,{}", base64::encode(png_bytes));
 
-                serde_json::to_value(&data_uri).context(error::SerdeJson)?
-            }
-        })
+            serde_json::to_value(&data_uri).context(error::SerdeJson)?
+        }
     };
 
-    let result = task_manager
-        .plot_executor()
-        .submit_ref(&workflow_id, task)
-        .await?;
+    let output = WrappedPlotOutput {
+        output_format,
+        plot_type,
+        data,
+    };
 
-    match result.as_ref() {
-        Ok(v) => Ok(web::Json(WrappedPlotOutput {
-            plot_type,
-            output_format,
-            data: v.clone(),
-        })),
-        Err(_) => Err(Error::NotYetImplemented),
-    }
+    Ok(web::Json(output))
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize)]
